requirements.txt
Django==2.2.6
djangorestframework==3.11.0
beautifulsoup4==4.4.1
requests==2.22.0
html5lib==0.999
django-cors-headers==3.2.0
